2019123 인수 첫번쨰 도전

+++ src/main/java/com/icia/moviefactory/entity/MovieReview.java
@@ -25,6 +25,6 @@ public class MovieReview {
 	private long repCnt;
 	private List<MovieReviewComment> comments;
 	private List<MovieReviewReport> reviewReports;
-	private long rating2;
+	private double rating2;
 }
